requirements.txt
fastapi
uvicorn
python-multipart
jinja2
numpy
opencv-python-headless
pillow
paddleocr
paddlepaddle
fuzzywuzzy
torch
